grouping sources in CodeSnippets
