+++ src/test/groovy/org/gradle/android/RoomWorkaroundAbstractTest.groovy
@@ -1,6 +1,5 @@
 package org.gradle.android
 
-import org.gradle.android.writer.JavaRoom
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.TaskOutcome
 
@@ -59,7 +58,7 @@ class RoomWorkaroundAbstractTest extends AbstractTest {
     }
 
     static void assertLegacySchemaUnchanged(File legacySchemaFile) {
-        assert legacySchemaFile.text == JavaRoom.legacySchemaJson()
+        assert legacySchemaFile.text == CodeSnippets.getRoomLegacySchemaJson()
     }
 
     void assertNotExecuted(buildResult, String taskPath) {
